+++ src/parser/util.ts
@@ -36,11 +36,6 @@ export function defrag<T extends Node, D extends Parser<unknown, any, S, C>[], S
   });
 }
 
-export function trimNode<P extends Parser<Node>>(parser: P): P;
-export function trimNode<T extends Node, D extends Parser<unknown, any, S, C>[], S extends object, C extends object>(parser: Parser<T, D, S, C>): Parser<T, D, S, C> {
-  return trimNode_(trimNode_(parser, 'start'), 'end');
-}
-
 export function trimNodeEnd<P extends Parser<Node>>(parser: P): P;
 export function trimNodeEnd<T extends Node, D extends Parser<unknown, any, S, C>[], S extends object, C extends object>(parser: Parser<T, D, S, C>): Parser<T, D, S, C> {
   return trimNode_(parser, 'end');