+++ src/parser/source/char.ts
@@ -3,6 +3,7 @@ import { Parser } from '../../combinator';
 import { text } from 'typed-dom';
 
 export function char(char: '#'): CharParser.SharpParser;
+export function char(char: '>'): CharParser.GreaterThanParser;
 export function char(char: '!'): CharParser.ExclamationParser;
 export function char(char: '='): CharParser.EqualParser;
 export function char(char: string): Parser<Text, []> {