+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.134.0
 
 - Change HTML parser to allow any size of leading and trailing whitespace.
+- Change HTML parser not to remove leading and trailing whitespace.
 - Fix HTML parser to parse entire invalid HTML syntax including closing tags.
 
 ## 0.133.1